+++ platform/common/eeprom.cpp
@@ -115,38 +115,7 @@ retry:
 	
 }
 
-struct eeprom_t {
-	#define	EE_HEADER	0xAA5533EE
-	u4_t header;                // 0x00
-	
-	#define	EE_FMT_REV	"A1"
-	char fmt_rev[2];            // 0x04
-	
-	char board_name[32];        // 0x06
-	char version[4];            // 0x26
-	char mfg[16];               // 0x2a
-	char part_no[16];           // 0x3a
-	char week[2];               // 0x4a
-	char year[2];               // 0x4c
-	char assembly[4];           // 0x4e
-	char serial_no[4];          // 0x52
-	
-	u2_t n_pins;                // 0x56
-	u2_t io_pins[EE_NPINS];     // 0x58
-	
-	#define	EE_MA_3V3	250
-	#define	EE_MA_5INT	0
-	#define	EE_MA_5EXT	0
-	#define	EE_MA_DC	1500
-	u2_t mA_3v3;                // 0xec
-	u2_t mA_5int;               // 0xee
-	u2_t mA_5ext;               // 0xf0
-	u2_t mA_DC;                 // 0xf2
-	
-	u1_t free[1];               // 0xf4+
-} __attribute__((packed));
-
-static eeprom_t eeprom;
+eeprom_t eeprom;
 static bool debian8 = false;
 
 #define EEPROM_DEV_DEBIAN7	    "/sys/bus/i2c/devices/1-0054/eeprom"
@@ -246,7 +215,7 @@ void eeprom_write(next_serno_e type, int serno)
 	SET_CHARS(e->part_no, "KIWISDR10", ' ');
 
 	// we use the WWYY fields as "date of last EEPROM write" rather than "date of production"
-	time_t t; time(&t);
+	time_t t = utc_time();
 	struct tm tm; gmtime_r(&t, &tm);
 	int ord_date = tm.tm_yday + 1;		// convert 0..365 to 1..366
 	int weekday = tm.tm_wday? tm.tm_wday : 7;	// convert Sunday = 0 to = 7
